requirements.txt
google-generativeai==0.8.5
python-dotenv==1.2.1
fastapi==0.123.0
uvicorn[standard]==0.38.0
